fix(injection): fix injection join
