+++ tube/etl/indexers/aggregation/new_translator.py
@@ -1,5 +1,5 @@
 from tube.etl.indexers.base.translator import Translator as BaseTranslator
-from tube.etl.indexers.base.prop import PropFactory, Prop
+from tube.etl.indexers.base.prop import PropFactory
 from tube.utils.general import (
     get_node_id_name,
     get_node_id_name_without_prefix,
@@ -7,8 +7,11 @@ from tube.utils.general import (
     PROJECT_CODE,
     PROGRAM_NAME,
 )
-from .parser import Parser
-from .nested.translator import Translator as NestedTranslator
+from tube.utils.spark import get_hadoop_type_ignore_fn
+from tube.etl.indexers.aggregation.parser import Parser
+from tube.etl.indexers.aggregation.nested.translator import (
+    Translator as NestedTranslator,
+)
 from tube.etl.indexers.base.logic import execute_filter
 from pyspark.sql.functions import (
     col,
@@ -439,9 +442,7 @@ class Translator(BaseTranslator):
                     )
                 else:
                     expr.append(
-                        lit(None)
-                        .cast(self.parser.get_hadoop_type_ignore_fn(prop))
-                        .alias(prop.name)
+                        lit(None).cast(get_hadoop_type_ignore_fn(prop)).alias(prop.name)
                     )
         df = df.groupBy(root_id).agg(*expr)
         return df
