+++ tube/etl/indexers/aggregation/nodes/direct_node.py
@@ -4,11 +4,18 @@ from tube.utils.dd import object_to_string
 
 class DirectNode(BaseNode):
     def __init__(
-        self, name, edge, props, sorted_by=None, desc_order=False, props_from_child=True
+        self,
+        name,
+        tbl_name,
+        edge,
+        props,
+        sorted_by=None,
+        desc_order=False,
+        props_from_child=True,
     ):
         super().__init__()
         self.name = name
-        self.tbl_name = name
+        self.tbl_name = tbl_name
         self.edge = edge
         self.props = props
         self.sorted_by = sorted_by
